sm_util, sm_at_commands: fix format-string and signed/unsigned defects
Addresses review findings M-1, M-2, M-3 and M-16.

* sm_at_commands.c: the "%.*s" precision argument must have type int per
  the C standard, but base_cmd_len[] is size_t. Add an explicit cast so
  AT#XCLAC output stays correct on builds where size_t is wider than int
  (CERT FIO47-C).
* sm_util.c: sm_util_at_printf() compared the possibly-negative vsnprintf()
  return value against sizeof(buf); the usual arithmetic conversions turn
  an encoding error into an apparent success. Check for < 0 first and
  return -EIO (CERT INT02-C, ERR33-C).
* sm_util.c: terminate_at_response() had the same int/size_t comparison
  plus an unguarded "len -= strlen(buf)" that could wrap. Guard both.
* sm_util.c: size_t values were printed with "%u"; use "%zu" (FIO47-C).
* sm_util.c: the scanf field widths passed to snprintf() are size_t but
  consumed as "%u"; cast to unsigned int explicitly.
* sm_util.c: size second_line[] with sizeof(literal) - 1 instead of
  strlen(), so it is a true constant expression rather than a VLA.
* sm_util.c: restoring the nrf_modem_at_scanf() trap macro aliased it to
  nrf_modem_at_printf, so any accidental later use would report the wrong
  alternative. Restore the original definition.

No functional change intended in the nominal paths.

+++ app/src/sm_at_commands.c
@@ -300,7 +300,12 @@ STATIC int handle_at_clac(enum at_parser_cmd_type cmd_type, struct at_parser *, 
 		}
 
 		if (!duplicate) {
-			rsp_send("%.*s\r\n", base_cmd_len[i],
+			/* The %.*s precision argument must have type int per the
+			 * C standard; base_cmd_len[] is size_t, so cast explicitly
+			 * rather than relying on size_t == unsigned int (true on
+			 * this 32-bit target, but not on a 64-bit host build).
+			 */
+			rsp_send("%.*s\r\n", (int)base_cmd_len[i],
 				 _nrf_modem_at_cmd_custom_list_start[i].cmd);
 		}
 	}
